requirements.txt
pymupdf>=1.24.0
Pillow>=10.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0

# LLM providers — install at least one:
groq>=0.11.0          # default (fastest + cheapest)
# openai>=1.30.0      # optional: for GPT-4o
# anthropic>=0.30.0   # optional: for Claude